Fix bug when the WeakValueMap is destructed before GC frees value
The problem was that the finalization callback did de-reference the
unordered_map instance kept as attribute map in WeakValueMap. This
was however deleted due to the WeakValueMap deletion.

Solution:
WeakValueMap instances are assigned a unique ID (64-bit) which is book-keeped
in a static unordered_map. It is added on WeakValueMap creation and erased on
deletion. The finalization callback checks if ID is present in unordered_map.