Defer WeakValueMap destructor until it is safe to use V8 API
Running the tests with a debug build of nodejs fails with:

  Fatal error in ../deps/v8/src/global-handles.cc, line 846
  Debug check failed: node_->state() == Node::NEAR_DEATH.

The root cause is that node's ObjectWrap invokes the destructor during a
first-pass weak reference callback.  The V8 API explicitly forbids
performing any V8 API calls at this time (other than resetting the
persistent handle).  In particular, the weak persistent handles of the
map's elements may be in the process of being reset, and manually
clearing them at that time results in a double-free of these handles.

The solution is to slightly modify ObjectWrap to defer the destructor
to a second-pass weak reference callback, at which time all persistent
handles are in a consistent state again.

This also fixes #3 and makes the complexity introduced by it
unnecessary.